+++ chaosazure/netapp/actions.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
+import logging
 import re
 
 from chaoslib.exceptions import FailedActivity
 from chaoslib.types import Configuration, Secrets
-from logzero import logger
 
 from chaosazure import init_netapp_management_client
 from chaosazure.common import cleanse
 from chaosazure.netapp.constants import RES_TYPE_SRV_NV
 from chaosazure.common.resources.graph import fetch_resources
+from chaosazure.vmss.records import Records
 
 __all__ = ["delete_netapp_volumes"]
-
-from chaosazure.vmss.records import Records
+logger = logging.getLogger("chaostoolkit")
 
 
 def delete_netapp_volumes(